Revert "➕ ADDS: Lists index page while refactoring existing component"
This reverts commit 7f20f2df03badf514d17ce674da77b61fdc35df7.
