+++ client/reader/index.ts
@@ -31,7 +31,6 @@ import {
 } from './controller';
 import postCacheMiddleware from './data/post/middleware';
 import {
-	allLists,
 	createList,
 	deleteList,
 	editList,
@@ -145,7 +144,6 @@ export default async function (): Promise< void > {
 	page( '/reader/feeds/lookup/*', redirectLoggedOutToSignup, feedLookup );
 
 	// Lists
-	page( '/reader/lists', redirectLoggedOut, sidebar, allLists, makeLayout, clientRender );
 	page( '/reader/list/:user/:list/edit/items', sidebar, editListItems, makeLayout, clientRender );
 	page( '/reader/list/:user/:list/edit', sidebar, editList, makeLayout, clientRender );
