+++ client/reader/user-profile/views/lists.scss
@@ -1,5 +1,5 @@
 .is-reader-page {
-	.reader-lists {
+	.user-profile__lists {
 		.summary-button {
 			margin-bottom: 16px;
 			padding: 18px 24px 24px;
